Moved log fields and context wrapper into a separated module
